+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostBotLikesPlugin/listeners/RatingChangedListener.kt
@@ -4,43 +4,37 @@ import com.github.insanusmokrassar.AutoPostBotLikesPlugin.database.LikesPluginLi
 import com.github.insanusmokrassar.AutoPostBotLikesPlugin.database.LikesPluginRegisteredLikesMessagesTable
 import com.github.insanusmokrassar.AutoPostBotLikesPlugin.models.ButtonMark
 import com.github.insanusmokrassar.AutoPostBotLikesPlugin.models.config.LikePluginConfig
+import com.github.insanusmokrassar.AutoPostBotLikesPlugin.utils.extensions.debounceByValue
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.commonLogger
-import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.executeAsync
-import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.subscribeChecking
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.*
 import com.pengrad.telegrambot.TelegramBot
 import com.pengrad.telegrambot.model.request.InlineKeyboardMarkup
 import com.pengrad.telegrambot.request.EditMessageReplyMarkup
+import kotlinx.coroutines.experimental.CancellationException
+import kotlinx.coroutines.experimental.channels.*
 import java.lang.ref.WeakReference
 
 class RatingChangedListener(
     private val likesPluginLikesTable: LikesPluginLikesTable,
     private val likesPluginRegisteredLikesMessagesTable: LikesPluginRegisteredLikesMessagesTable,
-    botWR: WeakReference<TelegramBot>,
+    private val botWR: WeakReference<TelegramBot>,
     private val chatId: Long,
     private val likePluginConfig: LikePluginConfig
 ) {
+    private val debounceDelay: Long = likePluginConfig.debounceDelay
+    private val retriesDelay = debounceDelay / 2
+
     init {
-        likesPluginLikesTable.messageButtonsUpdatedChannel.subscribeChecking {
-            val bot = botWR.get() ?: return@subscribeChecking false
-
-            updateMessage(bot, it)
-
-            true
+        likesPluginLikesTable.messageButtonsUpdatedChannel.debounceByValue(debounceDelay).subscribe {
+            updateMessage(it)
         }
-        likesPluginRegisteredLikesMessagesTable.messageIdAllocatedChannel.subscribeChecking {
-            val bot = botWR.get() ?: return@subscribeChecking false
-
-            updateMessage(bot, it)
-
-            true
+        likesPluginRegisteredLikesMessagesTable.messageIdAllocatedChannel.subscribe {
+            updateMessage(it)
         }
     }
 
-    private fun updateMessage(
-        bot: TelegramBot,
-        messageId: Int
-    ) {
-        bot.executeAsync(
+    private fun updateMessage(messageId: Int) {
+        botWR.get() ?.executeAsync(
             EditMessageReplyMarkup(
                 chatId,
                 messageId
@@ -51,7 +45,8 @@ class RatingChangedListener(
                 _, ioException ->
                 commonLogger.warning("Can't edit message $messageId for applying: ${ioException ?.message ?: "unknown problem"}")
             },
-            retries = 3
+            retries = 3,
+            retriesDelay = retriesDelay / 2
         )
     }